requirements.txt
bleak~=0.14.2
numpy<2.0
opencv-python<5.0
pillow>=10.4.0,<12.0